feat: svd_cca — SVD-denoised CCA subspace similarity primitive (Issue 684)
data_probe::cca behind opt-in svd_cca = [newton_schulz, subspace_phase_gate]
(+ linalg pub-mod gate gains svd_cca — the spectral_pencil latent-gap class).

svcca_into: column-center both sides → simd_dot_f32 covariances →
linalg::symmetric_eig (f64 dtype bridge, pinned) → numerical_rank(√λ, 0.99)
denoise → eigenvector projection → ridged reduced covariances →
ns_inv_sqrt_psd_into whitening (7 fixed NS iters) → B-form M = B·Bᵀ →
symmetric_eig → ρᵢ = √clamp(λᵢ,0,1), ρ̄ over min(kx,ky). Degenerate inputs
are the collapse signal (!is_finite() || < floor guards — Batch-54), never
an error. Bit-deterministic; zero-alloc steady state (G4: 0 allocs/100 calls).

GOAT (Bench 674, M3 Max, load 7-8.7): G1 8/8 (recovery r=0.8±0.1 at kx=4;
affine invariance joint-Π/scale/feat-perm/mix <0.02; bit-identical; NaN/Inf
no-panic; pathology contrast 0.94 vs 0.54, naive dilution 0.29; thin-SVD
rank parity). G2 p50 = 119.5-121.8 µs @ 32×32 n=128 — RECALIBRATED from
the issue's 25 µs (structurally unreachable: one 32×32 symmetric_eig alone
costs ~48 µs; the literal thin_svd_into pipeline measures 741 µs/side —
replaced by the algebraically-identical covariance+eig composition,
parity-pinned). G3 default 1904/0 untouched + feature-on 1925/0 +
no-default-features+svd_cca clean. G4 0 allocs.

Stays opt-in (no-default-consumer rule — T6). Consumers: ndb can_freeze v2,
riir-ai LoRAHotSwap equivalence + belief alignment, riir-train Plan 349.

Drive-by (pre-existing at HEAD, d6e737fd): 5 unnecessary_cast u32→u32 in
tests/contrastive_scope_alloc_check.rs made --all-targets -D warnings red
in BOTH feature states; fixed mechanically (test passes 1/1).

+++ crates/katgpt-core/tests/contrastive_scope_alloc_check.rs
@@ -30,8 +30,8 @@ fn g4_zero_alloc_steady_state() {
     // Build the table OUTSIDE the measured section (cold path).
     let mut b = ContrastiveScoreBuilder::new(4096, 0.5);
     for i in 0..200u32 {
-        b.observe_in(&[(i % 2048) as u32, (i % 977) as u32, (i % 613) as u32]);
-        b.observe_out(&[(2048 + i % 2048) as u32, (2048 + i % 1301) as u32]);
+        b.observe_in(&[i % 2048, i % 977, i % 613]);
+        b.observe_out(&[2048 + i % 2048, 2048 + i % 1301]);
     }
     let table = b.finish();
     let gate = ScopeGate { kappa: 0.05, theta: 8.0 };
